requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.5
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
drf-yasg==1.20.0
gunicorn==20.0.4
idna==3.2
importlib-metadata==4.6.4
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.10.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytz==2020.1
requests==2.26.0
ruamel.yaml==0.17.13
ruamel.yaml.clib==0.2.6
sqlparse==0.3.1
toml==0.10.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
zipp==3.5.0